pet-poultry-disease-classifier-main/pet-poultry-disease-classifier-main/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
huggingface_hub>=0.20.0
gunicorn>=21.0.0